+++ src/Predicate.ts
@@ -1,17 +1,18 @@
-import type { Unary } from "./Unary.js";
-
 /**
- * Unary function that returns a `boolean`.
+ * Unary function that returns a `boolean` and infers a given type for its
+ * argument.
  *
  * @category Function
  * @remarks
  * This type is useful for cases where a function needs to check if a certain
- * condition holds for an input value. For example, the type of a function that
- * checks if a number is even could be typed as `Predicate<number>`, a filtering
- * function could look like `Predicate<string | number, string>`, and so on.
+ * condition holds for an input value. For example, the type of a filtering
+ * function that filters strings in an array of strings and numbers could look
+ * like `Predicate<string | number, string>`.
  * @example
  * ```typescript
- * const isEven: Predicate<number> = number => number % 2 === 0;
+ * const isString: Predicate<number | string, string> = (
+ * 	numberOrString,
+ * ): numberOrString is string => typeof numberOrString === "string";
  * ```
  * @see {Unary}
  *
@@ -21,6 +22,6 @@ import type { Unary } from "./Unary.js";
  * @returns `true` if the predicate holds for the input value, and `false`
  * otherwise.
  */
-export type Predicate<Input, Predicated extends Input = Input> =
-	| Unary<Input, boolean>
-	| ((input: Input) => input is Predicated);
+export type Predicate<Input, Predicated extends Input = Input> = (
+	input: Input,
+) => input is Predicated;